+++ .env.sample
@@ -1,0 +1,5 @@
+ACCESS_KEY_ID=
+SECRET_ACCESS_KEY=
+REGION=""
+ENDPOINT=""
+PLANET_TABLE=""
