+++ package.json
@@ -4,7 +4,7 @@
   "description": "Demo en nodejs en arquitectura serverless",
   "main": "index.js",
   "scripts": {
-    "start": "serverless offline start --stage dev",
+    "start": "serverless offline start --stage demo",
     "test": "jest"
   },
   "keywords": [],
@@ -15,6 +15,7 @@
     "@aws-sdk/lib-dynamodb": "^3.614.0",
     "aws-serverless-express": "^3.4.0",
     "axios": "^1.7.2",
+    "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "serverless": "^4.1.11",
     "serverless-dynamodb-local": "^0.2.40",
@@ -25,6 +26,7 @@
     "yamljs": "^0.3.0"
   },
   "devDependencies": {
+    "serverless-dotenv-plugin": "^6.0.0",
     "serverless-middleware": "^3.3.0",
     "serverless-offline": "^13.6.0"
   }
